refactor(agents): share embedded skill preparation
Use one skill preparation owner for ordinary attempts and direct compaction while retaining their Code Mode, permission, prompt-budget and cleanup differences. Reuse prepared conversation facts without merging distinct agent identities.
