+++ src/agents/embedded-agent-runner/run/attempt-setup.test.ts
@@ -9,6 +9,7 @@ import { resolveSandboxContext as resolveRealSandboxContext } from "../../sandbo
 import type { SandboxContext } from "../../sandbox/types.js";
 import { castAgentMessage } from "../../test-helpers/agent-message-fixtures.js";
 import { createToolResultPromptProjectionState } from "../session-prompt-state.js";
+import { prepareEmbeddedSkills } from "../skill-runtime.js";
 import { buildEmbeddedForegroundPromptContext } from "./agent-end-context.js";
 import type { EmbeddedRunAttemptParams } from "./types.js";
 
@@ -26,7 +27,6 @@ vi.mock("../../sandbox.js", () => ({ resolveSandboxContext }));
 
 import {
   installEmbeddedAttemptContextGuards,
-  prepareEmbeddedAttemptSkills,
   prepareEmbeddedAttemptSetup,
   resolveAttemptWorkspaceSandbox,
 } from "./attempt-setup.js";
@@ -334,7 +334,7 @@ describe("prepareEmbeddedAttemptSetup", () => {
   });
 });
 
-describe("prepareEmbeddedAttemptSkills", () => {
+describe("prepareEmbeddedSkills", () => {
   it.each([
     { label: "unrestricted", toolExecutionAllow: undefined, readable: true },
     { label: "read allowed", toolExecutionAllow: ["read"], readable: true },
@@ -359,7 +359,8 @@ describe("prepareEmbeddedAttemptSkills", () => {
     await writeSkill(executionWorkspace, "execution-workspace-skill");
 
     try {
-      const prepared = prepareEmbeddedAttemptSkills({
+      const prepared = prepareEmbeddedSkills({
+        includeCodeModeSkills: true,
         attempt: {
           bootstrapWorkspaceDir: agentWorkspace,
           config: {},
