+++ src/agents/embedded-agent-runner/run/attempt-setup.ts
@@ -28,16 +28,9 @@ import {
   resolveProviderRuntimePluginHandle,
   type ProviderRuntimePluginHandle,
 } from "../../../plugins/provider-hook-runtime.js";
-import { resolveSkillsPrompt } from "../../../skills/loading/workspace-skill-prompt.js";
-import { resolveEmbeddedRunSkillEntries } from "../../../skills/runtime/embedded-run-entries.js";
-import {
-  applySkillEnvOverrides,
-  applySkillEnvOverridesFromSnapshot,
-} from "../../../skills/runtime/env-overrides.js";
 import { resolveUserPath } from "../../../utils.js";
 import { resolveSessionAgentIds } from "../../agent-scope.js";
 import { isHeartbeatLifecycleRunKind } from "../../bootstrap-mode.js";
-import { resolveCodeModeSkills, type CodeModeSkillReader } from "../../code-mode-skills.js";
 import { DEFAULT_CONTEXT_TOKENS } from "../../defaults.js";
 import type { EmbeddedContextFile } from "../../embedded-agent-helpers.js";
 import { resolveImageSanitizationLimits } from "../../image-sanitization.js";
@@ -46,16 +39,9 @@ import type { SandboxContext } from "../../sandbox/types.js";
 import type { guardSessionManager } from "../../session-tool-result-guard-wrapper.js";
 import { sanitizeToolUseResultPairingForModel } from "../../session-transcript-repair.js";
 import type { AgentSession } from "../../sessions/index.js";
-import { isToolExecutionAllowed } from "../../tool-policy-shared.js";
 import { invalidateComputerFrameIfMissing } from "../../tools/computer-tool.js";
 import { isCacheTtlEligibleProvider, readLastCacheTtlTimestamp } from "../cache-ttl.js";
 import { log } from "../logger.js";
-import {
-  createSandboxPromptEntryLoader,
-  mapSandboxSkillEntriesForPrompt,
-  mapSandboxSkillUsagePaths,
-  resolveSandboxSkillRuntimeInputs,
-} from "../sandbox-skills.js";
 import type { ToolResultPromptProjectionState } from "../session-prompt-state.js";
 import {
   installContextEngineLoopHook,
@@ -495,121 +481,6 @@ export function installEmbeddedAttemptContextGuards(input: {
       return request;
     },
   };
-}
-
-export function prepareEmbeddedAttemptSkills(params: {
-  attempt: EmbeddedRunAttemptParams;
-  effectiveWorkspace: string;
-  sandbox: EmbeddedAttemptSetup["sandbox"];
-  sessionAgentId: string;
-}) {
-  const executionAllow = params.attempt.toolExecutionAllow;
-  // Retained schemas are not execution permission. An unreadable skill catalog
-  // creates impossible prerequisites and exposes an ungated Code Mode reader.
-  if (
-    params.attempt.operation === "settled-tool-finalization" ||
-    (executionAllow && !isToolExecutionAllowed(executionAllow, "read"))
-  ) {
-    return {
-      restoreSkillEnv: () => {},
-      skillUsagePaths: undefined,
-      skillsPrompt: "",
-      skillsSnapshotForRun: undefined,
-      codeModeSkills: [],
-    };
-  }
-  const {
-    skillsEligibility,
-    skillsPromptWorkspaceDir,
-    skillsSnapshot,
-    skillsWorkspaceDir,
-    workspaceOnly,
-  } = resolveSandboxSkillRuntimeInputs({
-    sandbox: params.sandbox,
-    skillsAnchorWorkspace: params.attempt.bootstrapWorkspaceDir ?? params.effectiveWorkspace,
-    skillsSnapshot: params.attempt.skillsSnapshot,
-  });
-  const { shouldLoadSkillEntries, skillEntries, loadSkillEntries, preserveEntryOrder } =
-    resolveEmbeddedRunSkillEntries({
-      workspaceDir: skillsWorkspaceDir,
-      config: params.attempt.config,
-      agentId: params.sessionAgentId,
-      eligibility: skillsEligibility,
-      skillsSnapshot,
-      // Sandbox fallbacks stay inside their sandbox skill workspace;
-      // host execution skills are not mounted there.
-      ...(params.sandbox?.enabled === true
-        ? {}
-        : { executionSkillsDir: path.join(params.effectiveWorkspace, "skills") }),
-      workspaceOnly,
-    });
-  const restoreSkillEnv = skillsSnapshot
-    ? applySkillEnvOverridesFromSnapshot({
-        snapshot: skillsSnapshot,
-        config: params.attempt.config,
-      })
-    : applySkillEnvOverrides({
-        skills: skillEntries ?? [],
-        config: params.attempt.config,
-      });
-  try {
-    const promptSkillEntries = mapSandboxSkillEntriesForPrompt({
-      entries: shouldLoadSkillEntries ? skillEntries : undefined,
-      skillsWorkspaceDir,
-      skillsPromptWorkspaceDir,
-    });
-    const skillUsagePaths = mapSandboxSkillUsagePaths({
-      paths: params.sandbox?.skillUsagePaths,
-      skillsWorkspaceDir,
-      skillsPromptWorkspaceDir,
-    });
-    const skillsPrompt = resolveSkillsPrompt({
-      contextTokenBudget: params.attempt.contextTokenBudget,
-      skillsSnapshot,
-      entries: promptSkillEntries,
-      loadEntries: createSandboxPromptEntryLoader({
-        loadEntries: loadSkillEntries,
-        skillsWorkspaceDir,
-        skillsPromptWorkspaceDir,
-      }),
-      config: params.attempt.config,
-      workspaceDir: skillsPromptWorkspaceDir,
-      agentId: params.sessionAgentId,
-      eligibility: skillsEligibility,
-      preserveEntryOrder,
-    });
-    const sandbox = params.sandbox;
-    const sandboxSkillReader: CodeModeSkillReader | undefined = sandbox?.enabled
-      ? async ({ location, signal }) => {
-          const bridge = sandbox.fsBridge;
-          if (!bridge) {
-            throw new Error("Sandbox filesystem bridge is unavailable for skill reads.");
-          }
-          return (
-            await bridge.readFile({
-              filePath: location,
-              cwd: sandbox.containerWorkdir,
-              signal,
-            })
-          ).toString("utf8");
-        }
-      : undefined;
-    const codeModeSkills = resolveCodeModeSkills({
-      skillsPrompt,
-      candidates: skillsSnapshot?.resolvedSkills ?? skillEntries.map((entry) => entry.skill),
-      reader: sandboxSkillReader,
-    });
-    return {
-      restoreSkillEnv,
-      skillUsagePaths,
-      skillsPrompt,
-      skillsSnapshotForRun: skillsSnapshot,
-      codeModeSkills,
-    };
-  } catch (error) {
-    restoreSkillEnv();
-    throw error;
-  }
 }
 
 export type EmitDiagnosticRunCompleted = (
